+++ cmd/zdb/zdb.c
@@ -76,6 +76,7 @@
 #include <sys/btree.h>
 #include <zfs_comutil.h>
 #include <sys/zstd/zstd.h>
+#include <sys/zstd/zstd_impl.h>
 
 #include <libnvpair.h>
 #include <libzutil.h>
