+++ module/os/linux/zstd/Makefile.in
@@ -1,0 +1,7 @@
+#
+# Linux specific sources included from module/zstd/Makefile.in
+#
+
+ccflags-y += -I$(zstd_include)/..
+
+$(MODULE)-objs += ../os/linux/zstd/zstd_os.o
